refactor(#3721): campsite — one place states the per-column decode-failure policy
New row_decoder/column_decode_error module holds the rule (propagate; why no
class is skippable at these call sites) plus the ONE
end-of-partition-vs-defect decision the block loops share, so row_data.rs,
block_emit.rs and block_emit_windowed.rs all SHRINK. error.rs's inline tests
split to error_tests.rs (epic #1135) so the taxonomy file drops under 800.
